% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNNSIM-package.R
\docType{package}
\name{DNNSIM}
\alias{DNNSIM-package}
\alias{DNNSIM}
\title{The 'DNNSIM' package.}
\value{
This is the summary page. No return value.
}
\description{
Provides a deep neural network model with a monotonic increasing single index function tailored for periodontal disease studies. The residuals are assumed to follow a skewed T distribution, a skewed normal distribution, or a normal distribution. More details can be found at Liu, Huang, and Bai (2024) \url{doi:10.1016/j.csda.2024.108012}.
}
\author{
\strong{Maintainer}: Qingyang Liu \email{rh8liuqy@gmail.com} (\href{https://orcid.org/0000-0003-3265-6330}{ORCID})

Authors:
\itemize{
  \item Shijie Wang \email{shijiew.usc@gmail.com}
  \item Ray Bai \email{rbai@mailbox.sc.edu} (\href{https://orcid.org/0000-0002-7190-7844}{ORCID})
  \item Dipankar Bandyopadhyay \email{dbandyop@vcu.edu}
}

}
